+++ src/index.ts
@@ -3,12 +3,5 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import Schema from './Schema';
-import morph from './morph';
-import { MetadataField, PrimaryKey, Relation } from './types';
-
-export { PrimaryKey, Relation, MetadataField };
-
-export { morph };
-
-export default Schema;
+export { default } from './Schema';
+export { MetadataField, PrimaryKey, Relation } from './types';