+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * Can now generate and include multiple renderers (typed languages) in the output.
   * Added individual options `--flow`, `--prop-types`, and `--typescript`, to opt into generation.
 * Added polymorphic attribute support.
-  * `Schema`s can define polymorphic relations with `morphTo` and `morphToMany`.
+  * `Schema`s can define polymorphic relations with `morphTo`.
   * Schematics can define with the new "polymorph" field definition type.
 * Added proper `null` support for TypeScript.
 * Updated `WebpackPlugin` to support Webpack 4.